% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oda_pal.R
\name{oda_pal}
\alias{oda_pal}
\title{A wrapper for ggplot2::discrete scale which applies the CU Boulder ODA brand palette}
\usage{
oda_pal(primary = "Gold", other = "Black", discrete = TRUE)
}
\arguments{
\item{primary}{A string indicating the name of the Primary color to use (defaults to Gold)}

\item{other}{A string indicating the name of the Other color to use (defaults to Black)}

\item{discrete}{A boolean which determines if the returned palette is discrete or not. Defaults to true.}
}
\value{
A palette function corresponding to the CU Boulder ODA brand colors
}
\description{
A wrapper for ggplot2::discrete scale which applies the CU Boulder ODA brand palette
}
