% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_oda_gradient.R
\name{scale_color_oda_gradient}
\alias{scale_color_oda_gradient}
\title{A which applies the CU Boulder ODA brand palette as a ggplot gradient}
\usage{
scale_color_oda_gradient(low = "black", mid = NA, high = "gold", ...)
}
\arguments{
\item{low}{A string indicating the name of the color to use for low values of the gradient (defaults to black)}

\item{high}{A string indicating the name of the color to use for high values of the gradient (defaults to gold)}

\item{...}{Pass-through arguments for ggplot2::gradient or ggplot2::gradientn}

\item{med}{An optional argument specifying a third color for the mid-range values of the gradient (defaults to NA). If not provided a two-color gradient will be returned}
}
\value{
A ggplot2 gradient
}
\description{
A which applies the CU Boulder ODA brand palette as a ggplot gradient
}
