% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_oda.R
\name{scale_colour_oda}
\alias{scale_colour_oda}
\title{A wrapper for ggplot2::discrete scale which applies the CU Boulder ODA brand palette}
\usage{
scale_colour_oda(primary = "gold", other = "black", ...)
}
\arguments{
\item{primary}{A string indicating the name of the primary color to use (defaults to gold)}

\item{other}{A string indicating the name of the other color to use (defaults to black)}

\item{...}{Pass-through arguments for ggplot2::discrete_scale}
}
\value{
A ggplot2 discrete color scale
}
\description{
A wrapper for ggplot2::discrete scale which applies the CU Boulder ODA brand palette
}
