% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_oda.R
\name{theme_oda}
\alias{theme_oda}
\title{A basic CU Boulder branded ggplot2 theme for CU Boulder oda}
\usage{
theme_oda(base_size = 12, base_family = "sans", legend = TRUE)
}
\arguments{
\item{base_size}{The base size of all text in the plot. Title will be 1.5 times larger than this. Defaults to 12pt.}

\item{base_family}{A string indicating the base font family for all text in the plot. CU Boulder style guide indicates Helvetica should be used, which is the default (presuming the R `sans` font family has not been changed by the user)}

\item{legend}{A boolean indicating whether a legend should be included in the plot. Defaults to TRUE.}
}
\value{
A ggplot2 theme
}
\description{
A basic CU Boulder branded ggplot2 theme for CU Boulder oda
}
